Attach punctuation to words
The current regex means that punctuation at the end of a word is not attached to the word.  As a result, when the end of the word marking option is active, then the punctuation ends up separated from the syllable by the hyphen char: puer. -> pu-er-.
To fix this I modify the regex so that any punctuation characters which occur after a word (with no non-punctuation characters intervening) are attached to the word.  As a result the end of word hyphen char will not split the punctuation off from the word: puer. -> pu-er.-
This requires switching from re to regex, as the expanded escape classes which allow matching by Unicode category (https://www.regular-expressions.info/unicode.html#category) are not available in re.  This package is not necessarily available on all systems so perhaps this modification isn't for general distribution.